drop unused hosted service
